Added EmailTextWatcher class for android material edittext
